v1/backend/storage/repos/9f9d7705-da3e-45fb-b4fd-3b342766baf8/tuck-in-tales-backend/pyproject.toml
[project]
name = "tuck-in-tales-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Csaba Toth",email = "csaba@bitraptors.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "pillow>=11.3.0",
]

[tool.poetry]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
python-dotenv = {version = ">=1.1.0,<2.0.0"}
fastapi = {version = ">=0.115.12,<0.116.0"}
uvicorn = {version = ">=0.34.2,<0.35.0", extras = ["standard"]}
supabase = {version = ">=2.15.0,<3.0.0"}
pydantic-settings = {version = ">=2.9.1,<3.0.0"}
openai = {version = ">=1.75.0,<2.0.0"}
langchain = {version = ">=0.3.23,<0.4.0"}
langchain-openai = {version = ">=0.3.14,<0.4.0"}
firebase-admin = {version = ">=6.7.0,<7.0.0"}
python-multipart = {version = ">=0.0.20,<0.0.21"}
groq = "^0.22.0"
aiosqlite = "^0.21.0"
langgraph = "^0.3.34"
langgraph-checkpoint-sqlite = "^2.0.6"
google-generativeai = "^0.8.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
